(svn r2702) -Codechange: Cleaned up the sprite code and replaced many magic numbers by enums. There remains work in gfx.c to move the "transparency" and "recolor" bits around to make space for more sprites. However, 2800 additional sprites can now be loaded. There also remains cleanup and Doxygen work on many of the header files.

+++ openttd.h
@@ -71,7 +71,7 @@ typedef byte PlayerID;
 typedef byte OrderID;
 typedef byte CargoID;
 typedef uint16 StringID;
-typedef uint16 SpriteID;
+typedef uint32 SpriteID;
 typedef uint32 PalSpriteID;
 typedef uint32 CursorID;
 typedef uint16 EngineID; //! All enginenumbers should be of this type
